+++ src/StageKit.Fallout/PublishRid.cs
@@ -1,4 +1,5 @@
-﻿using StageKit.Primitives;
+﻿using System.Runtime.InteropServices;
+using StageKit.Primitives;
 
 namespace StageKit.Fallout;